Moved ClientError to its own file
